% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copperr_company.R
\name{cppr_createNewCompany}
\alias{cppr_createNewCompany}
\title{cppr_createNewCompany}
\usage{
cppr_createNewCompany(company)
}
\arguments{
\item{company}{list - a list of Company fields}
}
\value{
the id given to the new Company record if successful else NULL
}
\description{
Creates a new Company record in Copper using the list of values you provide.
You only need to provide the same mandatory fields required when creating a
Company manually in the Copper UI
}
\examples{
\dontrun{
fields <- list( `name` = "Kodak",
                `assignee_id` = 12345)
newCompanyId <- cppr_createNewCompany(fields)) 
if(!is.null(newCompanyId))
  sprint("successfully created a new Company with id \%d",newCompanyId)
}
}
\seealso{
\code{\link{cppr_getCompanyById}} for full list of available Company fields

Other Company functions: \code{\link{cppr_deleteCompany}},
  \code{\link{cppr_getCompanies}},
  \code{\link{cppr_getCompanyById}},
  \code{\link{cppr_getCompanyByName}},
  \code{\link{cppr_updateCompany}}
}
