% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copperr_leads.R
\name{cppr_createNewLead}
\alias{cppr_createNewLead}
\title{cppr_createNewLead}
\usage{
cppr_createNewLead(lead)
}
\arguments{
\item{lead}{list - containing at a minimum the mandatory fields of your new Lead}
}
\value{
the id of the Lead if created successfully, else NULL
}
\description{
Create a new Lead record using the fields passed as arguments
}
\examples{
\dontrun{ 
fields <- list( `name` = "New Lead",
                `assignee_id` = myUserId )
if(cppr_createNewLead(fields)) 
  sprintf("Successfully created new Lead \%d",id)
}
}
\seealso{
\code{\link{cppr_getLeads}} for full list of Lead fields

Other Lead functions: \code{\link{cppr_deleteLead}},
  \code{\link{cppr_getLeads}},
  \code{\link{cppr_updateLead}}
}
