% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copperr_opps.R
\name{cppr_createNewOpp}
\alias{cppr_createNewOpp}
\title{cppr_createNewOpp}
\usage{
cppr_createNewOpp(opp)
}
\arguments{
\item{opp}{list - a list of all the Opportunity fields}
}
\value{
the id of the new Opp record if successfully created, else NULL
}
\description{
Create a New Opportunity record in Copper using the field values passed in the argument
}
\examples{
\dontrun{ 
opp <- list( `name` = "New Name",
             `assignee_id` = 12345,
             `company_name` = "Kodak",
             `pipeline_id` = 23456)
newOppId <- cppr_updateOpp(34567,opp)
if(!is.null(newOppId)) 
  print("new opportunity successfully created with id \%d",newOppId)
}
}
\seealso{
\code{\link{cppr_getOppById}} for full list of Opportunity fields

Other Opportunity functions: \code{\link{cppr_deleteOpp}},
  \code{\link{cppr_getOppById}},
  \code{\link{cppr_getOppByName}},
  \code{\link{cppr_getOpps}},
  \code{\link{cppr_updateOppCloseDate}},
  \code{\link{cppr_updateOppCustomFieldDirect}},
  \code{\link{cppr_updateOppCustomFieldDropdownMulti}},
  \code{\link{cppr_updateOppCustomFieldDropdown}},
  \code{\link{cppr_updateOppStage}},
  \code{\link{cppr_updateOpp}}
}
