% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copperr_company.R
\name{cppr_deleteCompany}
\alias{cppr_deleteCompany}
\title{cppr_deleteCompany}
\usage{
cppr_deleteCompany(id)
}
\arguments{
\item{id}{numeric - the Id of the Company to be deleted}
}
\value{
the id of the Company record if successful else NULL
}
\description{
Permanently removes a Company record in Copper identified by the object id
}
\examples{
\dontrun{
if(cppr_deleteCompany(id))
  sprintf("succesfully deleted Company \%d",id)
}
}
\seealso{
\code{\link{cppr_getCompanyById}} for full list of Company fields

Other Company functions: \code{\link{cppr_createNewCompany}},
  \code{\link{cppr_getCompanies}},
  \code{\link{cppr_getCompanyById}},
  \code{\link{cppr_getCompanyByName}},
  \code{\link{cppr_updateCompany}}
}
