% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copperr_leads.R
\name{cppr_deleteLead}
\alias{cppr_deleteLead}
\title{cppr_deleteLead}
\usage{
cppr_deleteLead(id)
}
\arguments{
\item{id}{numeric - the Id of the Lead to be updated}
}
\value{
the id of the Lead if update was successful, else NULL
}
\description{
Permanently remove a Lead record identified by the the id passed 
as argument
}
\examples{
\dontrun{ 
if(!is.null(cppr_deleteLead(id))) 
  sprintf("Lead \%d successfully removed",id)
}
}
\seealso{
Other Lead functions: \code{\link{cppr_createNewLead}},
  \code{\link{cppr_getLeads}},
  \code{\link{cppr_updateLead}}
}
