% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copperr_opps.R
\name{cppr_deleteOpp}
\alias{cppr_deleteOpp}
\title{cppr_deleteOpp}
\usage{
cppr_deleteOpp(id)
}
\arguments{
\item{id}{numeric - the Id of the Opportunity to be updated}
}
\value{
the id of the opp if update was successful, else NULL
}
\description{
Permanently remove an Opportunity record identified by the the id passed 
as argument
}
\examples{
\dontrun{ 
if(!is.null(cppr_deleteOpp(id))) 
  sprintf("opportunity \%d successfully removed",id)
}
}
\seealso{
Other Opportunity functions: \code{\link{cppr_createNewOpp}},
  \code{\link{cppr_getOppById}},
  \code{\link{cppr_getOppByName}},
  \code{\link{cppr_getOpps}},
  \code{\link{cppr_updateCompanyCustomFieldDirect}},
  \code{\link{cppr_updateOppCloseDate}},
  \code{\link{cppr_updateOppCustomFieldDirect}},
  \code{\link{cppr_updateOppCustomFieldDropdownMulti}},
  \code{\link{cppr_updateOppCustomFieldDropdown}},
  \code{\link{cppr_updateOppStage}},
  \code{\link{cppr_updateOpp}},
  \code{\link{cppr_updatePersonCustomFieldDirect}}
}
