% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copperr_people.R
\name{cppr_deletePerson}
\alias{cppr_deletePerson}
\title{cppr_deletePerson}
\usage{
cppr_deletePerson(id)
}
\arguments{
\item{id}{numeric - the Id of the Person to be updated}
}
\value{
the id of the Person if update was successful, else NULL
}
\description{
Permanently remove an Person record identified by the the id passed 
as argument
}
\examples{
\dontrun{ 
if(!is.null(cppr_deletePerson(id))) 
  sprintf("Person \%d successfully removed",id)
}
}
\seealso{
For details of data structures refer to \href{https://developer.copper.com/#getting-support}{Copper (TM) Developer API Guide}

Other People functions: \code{\link{cppr_createNewPerson}},
  \code{\link{cppr_getAllRecordsRelatedToPerson}},
  \code{\link{cppr_getOppsRelatedToPerson}},
  \code{\link{cppr_getPeople}},
  \code{\link{cppr_getPersonByEmail}},
  \code{\link{cppr_getPersonById}},
  \code{\link{cppr_getPersonByName}},
  \code{\link{cppr_getPersonId}},
  \code{\link{cppr_updatePerson}}
}
