% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copper_resources.R
\name{cppr_extractCustomFieldValues}
\alias{cppr_extractCustomFieldValues}
\title{cppr_extractCustomFieldValues}
\usage{
cppr_extractCustomFieldValues(custom_fields, field_name)
}
\arguments{
\item{custom_fields}{data frame - set of custom field ids and values from a record 
returned from Copper using one of the cppr_getXxx functions}

\item{field_name}{string - the name text of the custom field, e.g. "Vertical"}
}
\value{
a list of values, or NULL if not found
}
\description{
Get a list of the value or values for a specific Custom Field of any Copper Object.
there may be multiple values e.g. if it is a MultiSelect field data type
}
\note{
on first call, stores list of custom field definitions in a global env 
to avoid unnecessary subsequent API calls
}
\examples{
\dontrun{ 
company <- cppr_getCompanyById(id)
values <- cppr_extractCustomFieldValues(company$custom_fields, "Industry")
print("there are",length(values),"values defined for this custom field:",values)
}
}
\seealso{
For details of data structures refer to \href{https://developer.copper.com/#getting-support}{Copper (TM) Developer API Guide}

Other Resource functions: \code{\link{cppr_getContactTypes}},
  \code{\link{cppr_getCustomFieldDatatype}},
  \code{\link{cppr_getCustomFieldDefinitions}},
  \code{\link{cppr_getCustomFieldId}},
  \code{\link{cppr_getCustomFieldName}},
  \code{\link{cppr_getCustomFieldOptionId}},
  \code{\link{cppr_getCustomFieldOptionName}},
  \code{\link{cppr_getCustomFieldOptionsAll}},
  \code{\link{cppr_getCustomerSourceId}},
  \code{\link{cppr_getCustomerSources}},
  \code{\link{cppr_getLeadStatusId}},
  \code{\link{cppr_getLeadStatuses}},
  \code{\link{cppr_getLossReasonId}},
  \code{\link{cppr_getLossReasons}},
  \code{\link{cppr_getPipelineStageId}},
  \code{\link{cppr_getPipelineStages}},
  \code{\link{cppr_getPipelines}}
}
