% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copperr_people.R
\name{cppr_getAllRecordsRelatedToPerson}
\alias{cppr_getAllRecordsRelatedToPerson}
\title{cppr_getAllRecordsRelatedToPerson}
\usage{
cppr_getAllRecordsRelatedToPerson(id)
}
\arguments{
\item{id}{string - the id of the person}
}
\value{
a data frame of record ids , or NULL on error.\cr
Field        Type    Details\cr
id           number  Unique identifier of the record\cr
type         string  the type of record, e.g. "opportunity", "company", "project"
}
\description{
Get the ids of all records in Copper related to a Person. These can be 
Opportunities, Companies, Projects, etc
}
\examples{
\dontrun{ 
relrecs <- cppr_getAllRecordsRelatedToPerson(12345)
relps <- dplyr::filter(relrcs,type=="project")
print(paste("this person is related to",nrow(relps),"projects"))
}
}
\seealso{
\code{\link{cppr_getOppsRelatedToPerson}}

\code{\link{cppr_getPersonId}}

For details of data structures refer to \href{https://developer.copper.com/#getting-support}{Copper (TM) Developer API Guide}

Other People functions: \code{\link{cppr_createNewPerson}},
  \code{\link{cppr_deletePerson}},
  \code{\link{cppr_getOppsRelatedToPerson}},
  \code{\link{cppr_getPeople}},
  \code{\link{cppr_getPersonByEmail}},
  \code{\link{cppr_getPersonById}},
  \code{\link{cppr_getPersonByName}},
  \code{\link{cppr_getPersonId}},
  \code{\link{cppr_updatePerson}}
}
