% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copper_resources.R
\name{cppr_getCustomerSourceId}
\alias{cppr_getCustomerSourceId}
\title{cppr_getCustomerSourceId}
\usage{
cppr_getCustomerSourceId(source_label)
}
\arguments{
\item{source_label}{- the name given to an option in the dropdown, e.g. "outbound"}
}
\value{
a numeric Id
}
\description{
get the Id for a named Customer Source (e.g. "Inbound")
}
\examples{
\dontrun{ 
id <- cppr_getCustomerSourceId("Inbound")
print(paste("the id for 'Inbound' source is", id))
}
}
\seealso{
Other Resource functions: \code{\link{cppr_extractCustomFieldValues}},
  \code{\link{cppr_getContactTypes}},
  \code{\link{cppr_getCustomFieldDatatype}},
  \code{\link{cppr_getCustomFieldDefinitions}},
  \code{\link{cppr_getCustomFieldId}},
  \code{\link{cppr_getCustomFieldName}},
  \code{\link{cppr_getCustomFieldOptionId}},
  \code{\link{cppr_getCustomFieldOptionName}},
  \code{\link{cppr_getCustomFieldOptionsAll}},
  \code{\link{cppr_getCustomerSources}},
  \code{\link{cppr_getLeadStatusId}},
  \code{\link{cppr_getLeadStatuses}},
  \code{\link{cppr_getLossReasonId}},
  \code{\link{cppr_getLossReasons}},
  \code{\link{cppr_getPipelineStageId}},
  \code{\link{cppr_getPipelineStages}},
  \code{\link{cppr_getPipelines}}
}
