% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copper_resources.R
\name{cppr_getCustomerSources}
\alias{cppr_getCustomerSources}
\title{cppr_getCustomerSources}
\usage{
cppr_getCustomerSources()
}
\value{
a data frame, or NULL on error
Field Type    Details
id    number  Unique identifier for the Contact Type.
name  string  The name of the Contact Type.
}
\description{
get a table of Customer Sources defined in Copper (e.g. "Inbound", etc)
for use in the corresponding field in Opportunity records
}
\examples{
\dontrun{ 
ct <- cppr_getCustomerSources()
print(paste("the customer sources defined in Copper are", ct$name))
}
}
\seealso{
Other Resource functions: \code{\link{cppr_getContactTypes}},
  \code{\link{cppr_getCustomFieldDatatype}},
  \code{\link{cppr_getCustomFieldDefinitions}},
  \code{\link{cppr_getCustomFieldId}},
  \code{\link{cppr_getCustomFieldName}},
  \code{\link{cppr_getCustomFieldOptionId}},
  \code{\link{cppr_getCustomFieldOptionName}},
  \code{\link{cppr_getCustomFieldOptionsAll}},
  \code{\link{cppr_getCustomerSourceId}},
  \code{\link{cppr_getLeadStatusId}},
  \code{\link{cppr_getLeadStatuses}},
  \code{\link{cppr_getLossReasonId}},
  \code{\link{cppr_getLossReasons}},
  \code{\link{cppr_getPipelineStageId}},
  \code{\link{cppr_getPipelineStages}},
  \code{\link{cppr_getPipelines}}
}
