% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copper_resources.R
\name{cppr_getCustomFieldDatatype}
\alias{cppr_getCustomFieldDatatype}
\title{cppr_getCustomFieldDatatype}
\usage{
cppr_getCustomFieldDatatype(field_name)
}
\arguments{
\item{field_name}{- the name text of the custom field, e.g. "Vertical"}
}
\value{
string - the datatype, or NULL if not found
data_type string enum   The type of data that should be stored within this custom field. 
                        Possible values are: String, Text, Dropdown, Date, Checkbox, Float, 
                        Url, Percentage, Currency
}
\description{
get the data type for a named custom field
}
\note{
on first call, stores list of custom field definitions in a global var 
to avoid unnecessary subsequent API calls
}
\examples{
\dontrun{ 
type <- cppr_getCustomFieldDatatype("Industry")
print(paste("the data type for the Industry custom field is",type))
}
}
\seealso{
Other Resource functions: \code{\link{cppr_getContactTypes}},
  \code{\link{cppr_getCustomFieldDefinitions}},
  \code{\link{cppr_getCustomFieldId}},
  \code{\link{cppr_getCustomFieldName}},
  \code{\link{cppr_getCustomFieldOptionId}},
  \code{\link{cppr_getCustomFieldOptionName}},
  \code{\link{cppr_getCustomFieldOptionsAll}},
  \code{\link{cppr_getCustomerSourceId}},
  \code{\link{cppr_getCustomerSources}},
  \code{\link{cppr_getLeadStatusId}},
  \code{\link{cppr_getLeadStatuses}},
  \code{\link{cppr_getLossReasonId}},
  \code{\link{cppr_getLossReasons}},
  \code{\link{cppr_getPipelineStageId}},
  \code{\link{cppr_getPipelineStages}},
  \code{\link{cppr_getPipelines}}
}
