% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copper_resources.R
\name{cppr_getCustomFieldDefinitions}
\alias{cppr_getCustomFieldDefinitions}
\title{cppr_getCustomFieldDefinitions}
\usage{
cppr_getCustomFieldDefinitions()
}
\value{
a data frame, or NULL if not found\cr
Field     Type          Details\cr
id        number        Unique identifier for the custom field definition.\cr
name      string        Label for the custom field definition\cr
data_type string enum   The type of data that should be stored within this custom field. 
                        Possible values are: String, Text, Dropdown, Date, Checkbox, Float, 
                        Url, Percentage, Currency\cr
currency  string enum   The currency used for this custom field definition. Valid only when \cr
                        the data type is Currency.\cr
options   options array A list of possible dropdown options. Valid only when the data type is Dropdown.
}
\description{
get a table of defintions for your custom fields
}
\examples{
\dontrun{ 
customFields <- cppr_getCustomFieldDefinitions()
print(paste("there are", nrow(customFields),"custom fields defined in Copper"))
}
}
\seealso{
For details of data structures refer to \href{https://developer.copper.com/#getting-support}{Copper (TM) Developer API Guide}

Other Resource functions: \code{\link{cppr_extractCustomFieldValues}},
  \code{\link{cppr_getContactTypes}},
  \code{\link{cppr_getCustomFieldDatatype}},
  \code{\link{cppr_getCustomFieldId}},
  \code{\link{cppr_getCustomFieldName}},
  \code{\link{cppr_getCustomFieldOptionId}},
  \code{\link{cppr_getCustomFieldOptionName}},
  \code{\link{cppr_getCustomFieldOptionsAll}},
  \code{\link{cppr_getCustomerSourceId}},
  \code{\link{cppr_getCustomerSources}},
  \code{\link{cppr_getLeadStatusId}},
  \code{\link{cppr_getLeadStatuses}},
  \code{\link{cppr_getLossReasonId}},
  \code{\link{cppr_getLossReasons}},
  \code{\link{cppr_getPipelineStageId}},
  \code{\link{cppr_getPipelineStages}},
  \code{\link{cppr_getPipelines}}
}
