% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copper_resources.R
\name{cppr_getCustomFieldId}
\alias{cppr_getCustomFieldId}
\title{cppr_getCustomFieldId}
\usage{
cppr_getCustomFieldId(field_name)
}
\arguments{
\item{field_name}{- the name text of the custom field, e.g. "Vertical"}
}
\value{
the Id of that custom field, or NULL if not found
}
\description{
get the Id of a named custom fields
}
\note{
on the first call to this function, a list of custom field definitions is 
stored in a local environment for performance - to avoid unnecessary subsequent API calls
}
\examples{
\dontrun{ 
id <- cppr_getCustomFieldId("Industry")
print(paste("the Id for the Industry custom field is",id))
}
}
\seealso{
For details of data structures refer to \href{https://developer.copper.com/#getting-support}{Copper (TM) Developer API Guide}

Other Resource functions: \code{\link{cppr_extractCustomFieldValues}},
  \code{\link{cppr_getContactTypes}},
  \code{\link{cppr_getCustomFieldDatatype}},
  \code{\link{cppr_getCustomFieldDefinitions}},
  \code{\link{cppr_getCustomFieldName}},
  \code{\link{cppr_getCustomFieldOptionId}},
  \code{\link{cppr_getCustomFieldOptionName}},
  \code{\link{cppr_getCustomFieldOptionsAll}},
  \code{\link{cppr_getCustomerSourceId}},
  \code{\link{cppr_getCustomerSources}},
  \code{\link{cppr_getLeadStatusId}},
  \code{\link{cppr_getLeadStatuses}},
  \code{\link{cppr_getLossReasonId}},
  \code{\link{cppr_getLossReasons}},
  \code{\link{cppr_getPipelineStageId}},
  \code{\link{cppr_getPipelineStages}},
  \code{\link{cppr_getPipelines}}
}
