% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copper_resources.R
\name{cppr_getCustomFieldName}
\alias{cppr_getCustomFieldName}
\title{cppr_getCustomFieldName}
\usage{
cppr_getCustomFieldName(field_id)
}
\arguments{
\item{field_id}{- the idof the custom field, e.g. 127800}
}
\value{
the Id of that custom field, or NULL if not found
}
\description{
get the name of a Custom Field by Id
}
\note{
stores list of custom field definitions in a global env 
to avoid unnecessary subsequent API calls
}
\examples{
\dontrun{ 
name <- cppr_getCustomFieldName(20278)
print(name)
}
}
\seealso{
For details of data structures refer to \href{https://developer.copper.com/#getting-support}{Copper (TM) Developer API Guide}

Other Resource functions: \code{\link{cppr_extractCustomFieldValues}},
  \code{\link{cppr_getContactTypes}},
  \code{\link{cppr_getCustomFieldDatatype}},
  \code{\link{cppr_getCustomFieldDefinitions}},
  \code{\link{cppr_getCustomFieldId}},
  \code{\link{cppr_getCustomFieldOptionId}},
  \code{\link{cppr_getCustomFieldOptionName}},
  \code{\link{cppr_getCustomFieldOptionsAll}},
  \code{\link{cppr_getCustomerSourceId}},
  \code{\link{cppr_getCustomerSources}},
  \code{\link{cppr_getLeadStatusId}},
  \code{\link{cppr_getLeadStatuses}},
  \code{\link{cppr_getLossReasonId}},
  \code{\link{cppr_getLossReasons}},
  \code{\link{cppr_getPipelineStageId}},
  \code{\link{cppr_getPipelineStages}},
  \code{\link{cppr_getPipelines}}
}
