% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copper_resources.R
\name{cppr_getCustomFieldOptionId}
\alias{cppr_getCustomFieldOptionId}
\title{cppr_getCustomFieldOptionId}
\usage{
cppr_getCustomFieldOptionId(field_name, option_label)
}
\arguments{
\item{field_name}{- the name text of the custom field, e.g. "Vertical"}

\item{option_label}{- the label text of the option , e.g. "Transport"}
}
\value{
the Id number of the option label, or NULL if not found
}
\description{
get the Id of a named Custom Field Option for a custom field of dropdown or 
MultiSelect type.
Each dropdown option you define in Copper for a custom field is given an Id by Copper
}
\note{
on first call, stores list of custom field definitions in a global env 
to avoid unnecessary subsequent API calls
}
\examples{
\dontrun{ 
id <- cppr_getCustomFieldOptionId("Industry","Finance")
}
}
\seealso{
For details of data structures refer to \href{https://developer.copper.com/#getting-support}{Copper (TM) Developer API Guide}

Other Resource functions: \code{\link{cppr_extractCustomFieldValues}},
  \code{\link{cppr_getContactTypes}},
  \code{\link{cppr_getCustomFieldDatatype}},
  \code{\link{cppr_getCustomFieldDefinitions}},
  \code{\link{cppr_getCustomFieldId}},
  \code{\link{cppr_getCustomFieldName}},
  \code{\link{cppr_getCustomFieldOptionName}},
  \code{\link{cppr_getCustomFieldOptionsAll}},
  \code{\link{cppr_getCustomerSourceId}},
  \code{\link{cppr_getCustomerSources}},
  \code{\link{cppr_getLeadStatusId}},
  \code{\link{cppr_getLeadStatuses}},
  \code{\link{cppr_getLossReasonId}},
  \code{\link{cppr_getLossReasons}},
  \code{\link{cppr_getPipelineStageId}},
  \code{\link{cppr_getPipelineStages}},
  \code{\link{cppr_getPipelines}}
}
