% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copper_resources.R
\name{cppr_getCustomFieldOptionName}
\alias{cppr_getCustomFieldOptionName}
\title{cppr_getCustomFieldOptionName}
\usage{
cppr_getCustomFieldOptionName(field_name, option_id)
}
\arguments{
\item{field_name}{- the name text of the custom field, e.g. "Vertical"}

\item{option_id}{- the Id of the option , e.g. 230189}
}
\value{
the text label for the option, or NULL if not found
}
\description{
get the name of a Custom Field Option by Id, for a custom field of dropdown type.
Each dropdown option you define in Copper for a custom field is given an Id by Copper
}
\note{
on first call, stores list of custom field definitions in a global env 
to avoid unnecessary subsequent API calls
}
\examples{
\dontrun{ 
name <- cppr_getCustomFieldOptionName("Industry",202193)
print(paste("The dropdown option 202193 is displayed as",name)
}
}
\seealso{
Other Resource functions: \code{\link{cppr_getContactTypes}},
  \code{\link{cppr_getCustomFieldDatatype}},
  \code{\link{cppr_getCustomFieldDefinitions}},
  \code{\link{cppr_getCustomFieldId}},
  \code{\link{cppr_getCustomFieldName}},
  \code{\link{cppr_getCustomFieldOptionId}},
  \code{\link{cppr_getCustomFieldOptionsAll}},
  \code{\link{cppr_getCustomerSourceId}},
  \code{\link{cppr_getCustomerSources}},
  \code{\link{cppr_getLeadStatusId}},
  \code{\link{cppr_getLeadStatuses}},
  \code{\link{cppr_getLossReasonId}},
  \code{\link{cppr_getLossReasons}},
  \code{\link{cppr_getPipelineStageId}},
  \code{\link{cppr_getPipelineStages}},
  \code{\link{cppr_getPipelines}}
}
