% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copperr_leads.R
\name{cppr_getLeads}
\alias{cppr_getLeads}
\title{cppr_getLeads}
\usage{
cppr_getLeads(assignee_id = NULL, country = NULL, state = NULL,
  tags = NULL, sort_by = "name", sort_direction = "asc")
}
\arguments{
\item{assignee_id}{numeric - select leads assigned to this User}

\item{country}{string - select leads from this country (specify the 2-letter country code, e.g. "NZ")}

\item{state}{string - select leads from this state (specify the full name of state, e.g. "Idaho")}

\item{tags}{string - select leads with one of these tags}

\item{sort_by}{string - "name", etc}

\item{sort_direction}{- allowed values "asc" or "desc"}
}
\value{
a data frame
Field          Type    Details
id             number  Unique identifier 
name           string  firstname lastname
address        data frame encapulsation of street,city,postcode,state, country
assignee_id    number  Unique identifier of the User that will be the owner of the Lead.
company_id     string  The unique identifier of the primary Company with which the Lead is associated.
company_name   string  The name of the primary Company with which the Lead is associated.
customer_source_id number  The unique identifier of the source of the Lead
details        string  Description of the Lead.
email[]       data frame  An array of email addresses belonging to the Lead.
email[].email  string  An email address.
email[].category string  The category of the email address.
phone_numbers[] list  An array of phone numbers belonging to the Lead.
phone_numbers[].number  string  A phone number.
phone_numbers[].category  string  The category of the phone number.
socials[]       list  An array of social profiles belonging to the Lead.
socials[].url    string  The URL of a social profile.
socials[].category  string  The category of the social profile.
status         string_enum 
tags           list  An array of the tags associated with the Lead, represented as strings.
title          string  The professional title of the Lead.
websites[]     list  An array of websites belonging to the Lead.
websites[].url  string  The URL of a website.
websites[].category string  The category of the website.
date_created   number  A Unix timestamp representing the time at which this Lead was created.
date_modified  number  A Unix timestamp representing the time at which this Lead was last modified.
custom_fields[] list  An array of custom field values belonging to the Person.
custom_fields[]   
.custom_field_definition_id number  The id of the Custom Field Definition for which this Custom Field stores a value.
custom_fields[]   
.value  mixed The value (number, string, option id, or timestamp) of this Custom Field.
}
\description{
Get a table containg all Leads from Copper. You can filter the recirds using the optional arguments.
Each row of the table is a Lead, represented by a list of field names and values. 
Some fields are themselves lists.
}
\examples{
\dontrun{ 
leads <- cppr_getLeads()
print(paste("retrieved",nrow(leads),"leads from PW"))
}
}
\seealso{
Other Lead functions: \code{\link{cppr_createNewLead}},
  \code{\link{cppr_deleteLead}},
  \code{\link{cppr_updateLead}}
}
