% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copper_resources.R
\name{cppr_getLeadStatusId}
\alias{cppr_getLeadStatusId}
\title{cppr_getLeadStatusId}
\usage{
cppr_getLeadStatusId(name)
}
\arguments{
\item{name}{string - the label (name) of the status, e.g. "Qualified"}
}
\value{
an numeric Id. The first matching Copper Id, or NULL if not found.
}
\description{
Get a Copper Id for the Status  identfied by the name argument
}
\examples{
\dontrun{ 
pipes <- cppr_getLeadStatuses() 
id <- cppr_getLeadStatusId(pipes$id[1],"Qualified")
if(!is.null(id)) 
  print(paste("stage 'Qualified' has the id",id))
}
}
\seealso{
Other Resource functions: \code{\link{cppr_getContactTypes}},
  \code{\link{cppr_getCustomFieldDatatype}},
  \code{\link{cppr_getCustomFieldDefinitions}},
  \code{\link{cppr_getCustomFieldId}},
  \code{\link{cppr_getCustomFieldName}},
  \code{\link{cppr_getCustomFieldOptionId}},
  \code{\link{cppr_getCustomFieldOptionName}},
  \code{\link{cppr_getCustomFieldOptionsAll}},
  \code{\link{cppr_getCustomerSourceId}},
  \code{\link{cppr_getCustomerSources}},
  \code{\link{cppr_getLeadStatuses}},
  \code{\link{cppr_getLossReasonId}},
  \code{\link{cppr_getLossReasons}},
  \code{\link{cppr_getPipelineStageId}},
  \code{\link{cppr_getPipelineStages}},
  \code{\link{cppr_getPipelines}}
}
