% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copper_resources.R
\name{cppr_getLossReasonId}
\alias{cppr_getLossReasonId}
\title{cppr_getLossReasonId}
\usage{
cppr_getLossReasonId(name)
}
\arguments{
\item{name}{string - the label (name) of the status, e.g. "Budget"}
}
\value{
an numeric Id. The first matching Copper Id, or NULL if not found.
}
\description{
Get a Copper Id for the Loss Reason identfied by the name argument
}
\examples{
\dontrun{ 
pipes <- cppr_getLossReasons() 
id <- cppr_getLossReasonId(pipes$id[1],"Budget")
if(!is.null(id)) 
  print(paste("stage 'Budget' has the id",id))
}
}
\seealso{
Other Resource functions: \code{\link{cppr_getContactTypes}},
  \code{\link{cppr_getCustomFieldDatatype}},
  \code{\link{cppr_getCustomFieldDefinitions}},
  \code{\link{cppr_getCustomFieldId}},
  \code{\link{cppr_getCustomFieldName}},
  \code{\link{cppr_getCustomFieldOptionId}},
  \code{\link{cppr_getCustomFieldOptionName}},
  \code{\link{cppr_getCustomFieldOptionsAll}},
  \code{\link{cppr_getCustomerSourceId}},
  \code{\link{cppr_getCustomerSources}},
  \code{\link{cppr_getLeadStatusId}},
  \code{\link{cppr_getLeadStatuses}},
  \code{\link{cppr_getLossReasons}},
  \code{\link{cppr_getPipelineStageId}},
  \code{\link{cppr_getPipelineStages}},
  \code{\link{cppr_getPipelines}}
}
