% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copper_resources.R
\name{cppr_getLossReasons}
\alias{cppr_getLossReasons}
\title{cppr_getLossReasons}
\usage{
cppr_getLossReasons()
}
\value{
a data frame, or NULL on error\cr
Field                     Details\cr
id (number)               Unique identifier for the Pipeline Stage.\cr
name (string)             The name of the Pipeline Stage.\cr
order (number)            The position it is displayed in the list\cr
is_default (boolean)      indicates if this value is the default for new records
}
\description{
get a table containing the definitions of the Loss Reasons defined
for use in the corresponding field in Opportunity records
}
\examples{
\dontrun{ 
statuses <- cppr_getLossReasons()
print(paste("the statuses are called",statuses$name))
}
}
\seealso{
Other Resource functions: \code{\link{cppr_extractCustomFieldValues}},
  \code{\link{cppr_getContactTypes}},
  \code{\link{cppr_getCustomFieldDatatype}},
  \code{\link{cppr_getCustomFieldDefinitions}},
  \code{\link{cppr_getCustomFieldId}},
  \code{\link{cppr_getCustomFieldName}},
  \code{\link{cppr_getCustomFieldOptionId}},
  \code{\link{cppr_getCustomFieldOptionName}},
  \code{\link{cppr_getCustomFieldOptionsAll}},
  \code{\link{cppr_getCustomerSourceId}},
  \code{\link{cppr_getCustomerSources}},
  \code{\link{cppr_getLeadStatusId}},
  \code{\link{cppr_getLeadStatuses}},
  \code{\link{cppr_getLossReasonId}},
  \code{\link{cppr_getPipelineStageId}},
  \code{\link{cppr_getPipelineStages}},
  \code{\link{cppr_getPipelines}}
}
