% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copperr_opps.R
\name{cppr_getOppByName}
\alias{cppr_getOppByName}
\title{cppr_getOppByName}
\usage{
cppr_getOppByName(name)
}
\arguments{
\item{name}{string - the full name of the opportunity}
}
\value{
the list of Opportunity fields, or Null on failure
}
\description{
Get an Opportunity record identified by its Name.
A record is a list of field names and values
}
\note{
This uses the search API, which could return multiple opportunities with the same name. 
Only the first match is returned! (if you initialized the package with Verbose=TRUE, 
you will see a warning displayed in this case)
To be sure of getting the record you want, use cppr_getOppById which uses
a field guaranteed by Copper to be unique, or use cppr_getOpps() to get all matches.
}
\examples{
\dontrun{ 
opp <- cppr_getOppByName("Kodak film")
}
}
\seealso{
\code{\link{cppr_getOppById}} for full list of Opportunity fields

For details of data structures refer to \href{https://developer.copper.com/#getting-support}{Copper (TM) Developer API Guide}

Other Opportunity functions: \code{\link{cppr_createNewOpp}},
  \code{\link{cppr_deleteOpp}},
  \code{\link{cppr_getOppById}},
  \code{\link{cppr_getOpps}},
  \code{\link{cppr_updateOppCloseDate}},
  \code{\link{cppr_updateOppCustomFieldDirect}},
  \code{\link{cppr_updateOppCustomFieldDropdownMulti}},
  \code{\link{cppr_updateOppCustomFieldDropdown}},
  \code{\link{cppr_updateOppStage}},
  \code{\link{cppr_updateOpp}}
}
