% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copperr_opps.R
\name{cppr_getOpps}
\alias{cppr_getOpps}
\title{cppr_getOpps
Get a table containing all Opportunities.
Each row of the data frame returned contains a list of field names and values.}
\usage{
cppr_getOpps(pipeline_id = NULL, pipeline_stage_id = NULL,
  assignee_id = NULL, tags = NULL, sort_by = "name",
  sort_direction = "asc")
}
\arguments{
\item{pipeline_id}{numeric - select opps in this pipeline}

\item{pipeline_stage_id}{numeric - select opps at this stage in the pipeline}

\item{assignee_id}{numeric - select opps assigned to this User}

\item{tags}{string - select leads with one of these tags}

\item{sort_by}{string - values can be any of : (can be combined to provided ANDed)
\itemize{
  \item name
  \item assignee
  \item company_name
  \item customer_source_id
  \item monetary_value
  \item primary_contact
  \item priority
  \item status
  \item inactive_days
  \item last_interaction
}}

\item{sort_direction}{string - can be "desc" or "asc"}
}
\value{
a dataframe of Opportunity field lists, or Null if no matching Opps found
}
\description{
cppr_getOpps
Get a table containing all Opportunities.
Each row of the data frame returned contains a list of field names and values.
}
\examples{
\dontrun{ 
opps <- cppr_getOpps()
print(paste("retrieved",nrow(opps),"opportunities from Copper"))
}
}
\seealso{
\code{\link{cppr_getOppById}} for full list of Opportunity fields

For details of data structures refer to \href{https://developer.copper.com/#getting-support}{Copper (TM) Developer API Guide}

Other Opportunity functions: \code{\link{cppr_createNewOpp}},
  \code{\link{cppr_deleteOpp}},
  \code{\link{cppr_getOppById}},
  \code{\link{cppr_getOppByName}},
  \code{\link{cppr_updateCompanyCustomFieldDirect}},
  \code{\link{cppr_updateOppCloseDate}},
  \code{\link{cppr_updateOppCustomFieldDirect}},
  \code{\link{cppr_updateOppCustomFieldDropdownMulti}},
  \code{\link{cppr_updateOppCustomFieldDropdown}},
  \code{\link{cppr_updateOppStage}},
  \code{\link{cppr_updateOpp}},
  \code{\link{cppr_updatePersonCustomFieldDirect}}
}
