% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copperr_people.R
\name{cppr_getPersonById}
\alias{cppr_getPersonById}
\title{cppr_getPersonById}
\usage{
cppr_getPersonById(id)
}
\arguments{
\item{id}{numeric - the Copper Id of the person}
}
\value{
a list, or NULL if unsuccessful.\cr
Field        Type    Details\cr
id           number  Unique identifier.\cr
name*        string  firstname lastname.\cr
address      address An encapsulation of the Person's street, city, state, postal code, and country.\cr
assignee_id  number Unique identifier of the User that will be the owner of the Person.\cr
etc
}
\description{
Get a People record identfied by the person's Id
}
\examples{
\dontrun{ 
id <- cppr_getPersonId("John Smith")
if( nrow(id) == 1 ) # found exactly one matching record in Copper
  person <- cppr_getPersonById(id)
}
}
\seealso{
\code{\link{cppr_getPeople}} for full list of Person fields

\code{\link{cppr_getPersonId}} to find the Person Id if you know their name

For details of data structures refer to \href{https://developer.copper.com/#getting-support}{Copper (TM) Developer API Guide}

Other People functions: \code{\link{cppr_createNewPerson}},
  \code{\link{cppr_deletePerson}},
  \code{\link{cppr_getAllRecordsRelatedToPerson}},
  \code{\link{cppr_getOppsRelatedToPerson}},
  \code{\link{cppr_getPeople}},
  \code{\link{cppr_getPersonByEmail}},
  \code{\link{cppr_getPersonByName}},
  \code{\link{cppr_getPersonId}},
  \code{\link{cppr_updatePerson}}
}
