% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copperr_people.R
\name{cppr_getPersonByName}
\alias{cppr_getPersonByName}
\title{cppr_getPersonByName}
\usage{
cppr_getPersonByName(name)
}
\arguments{
\item{name}{string - the full name of the person 'firstname lastname'}
}
\value{
a list of Person fields, or NULL if unsuccessful.\cr
Field        Type    Details\cr
id           number  Unique identifier\cr
name         string  firstname lastname.\cr
address      address An encapsulation of the Person's street, city, state, postal code, and country.\cr
assignee_id  number Unique identifier of the User that will be the owner of the Person.\cr
etc
}
\description{
Get a People record identfied by the person's Name
}
\note{
This uses the search API, which could return multiple people with the same name. 
Only the first match is returned! (if you initialized the package with Verbose=TRUE, 
you will see a warning displayed in this case)
To be sure of getting the record you want, use cppr_getPersonById or cppr_getPersonByEmail which both
use a field guaranteed by Copper to be unique, or use cppr_getPeople() to get all matches.
}
\examples{
\dontrun{ 
person <- cppr_getPersonByName("John Smith")
print(paste("Person is related to the company",person$company_name))
}
}
\seealso{
\code{\link{cppr_getPeople}} for full list of Person fields

For details of data structures refer to \href{https://developer.copper.com/#getting-support}{Copper (TM) Developer API Guide}

Other People functions: \code{\link{cppr_createNewPerson}},
  \code{\link{cppr_deletePerson}},
  \code{\link{cppr_getAllRecordsRelatedToPerson}},
  \code{\link{cppr_getOppsRelatedToPerson}},
  \code{\link{cppr_getPeople}},
  \code{\link{cppr_getPersonByEmail}},
  \code{\link{cppr_getPersonById}},
  \code{\link{cppr_getPersonId}},
  \code{\link{cppr_updatePerson}}
}
