% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copperr_people.R
\name{cppr_getPersonId}
\alias{cppr_getPersonId}
\title{cppr_getPersonId}
\usage{
cppr_getPersonId(name)
}
\arguments{
\item{name}{string - the full name of the person in format 'firstname lastname'}
}
\value{
an numeric Id. The first matching Copper Id, 0 if no match found, or NULL on error.
}
\description{
Get a Copper Id for the Person record identfied by the person's Name
}
\note{
This uses the search API, which could return multiple people with the same name. 
Only the first match is returned! (if you initialized the package with Verbose=TRUE, 
you will see a warning displayed in this case)
To be sure of getting the record you want, use cppr_getPersonById or cppr_getPersonByEmail which both
use a field guaranteed by Copper to be unique, or use getUsers() to get all matches.
}
\examples{
\dontrun{ 
id <- cppr_getPersonId("John Smith")
print(paste("found",nrow(id),"matching People with this name in Copper"))
}
}
\seealso{
Other People functions: \code{\link{cppr_createNewPerson}},
  \code{\link{cppr_deletePerson}},
  \code{\link{cppr_getAllRecordsRelatedToPerson}},
  \code{\link{cppr_getOppsRelatedToPerson}},
  \code{\link{cppr_getPeople}},
  \code{\link{cppr_getPersonByEmail}},
  \code{\link{cppr_getPersonById}},
  \code{\link{cppr_getPersonByName}},
  \code{\link{cppr_updatePerson}}
}
