% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copper_resources.R
\name{cppr_getPipelines}
\alias{cppr_getPipelines}
\title{cppr_getPipelines}
\usage{
cppr_getPipelines()
}
\value{
a data frame, or NULL if none found\cr
Field                     Details\cr
id (number)               Unique identifier for the Pipeline\cr
name (string)             The name of the Pipeline.\cr
stages (list)             The Stages in this pipeline.\cr
win_probability (number)  The expected probability of winning an Opportunity in this Pipeline Stage. Valid values are [0-100] (inclusive).
}
\description{
get a table containing the definitions of the stages defined for a specified Pipeline
Opportunity records can be assigned to a Pipeline in Copper
}
\examples{
\dontrun{ 
pipelines <- cppr_getPipelines()
print(paste("the pipeline Ids are ",pipelines$name))
}
}
\seealso{
For details of data structures refer to \href{https://developer.copper.com/#getting-support}{Copper (TM) Developer API Guide}

Other Resource functions: \code{\link{cppr_extractCustomFieldValues}},
  \code{\link{cppr_getContactTypes}},
  \code{\link{cppr_getCustomFieldDatatype}},
  \code{\link{cppr_getCustomFieldDefinitions}},
  \code{\link{cppr_getCustomFieldId}},
  \code{\link{cppr_getCustomFieldName}},
  \code{\link{cppr_getCustomFieldOptionId}},
  \code{\link{cppr_getCustomFieldOptionName}},
  \code{\link{cppr_getCustomFieldOptionsAll}},
  \code{\link{cppr_getCustomerSourceId}},
  \code{\link{cppr_getCustomerSources}},
  \code{\link{cppr_getLeadStatusId}},
  \code{\link{cppr_getLeadStatuses}},
  \code{\link{cppr_getLossReasonId}},
  \code{\link{cppr_getLossReasons}},
  \code{\link{cppr_getPipelineStageId}},
  \code{\link{cppr_getPipelineStages}}
}
