% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copper_resources.R
\name{cppr_getPipelineStageId}
\alias{cppr_getPipelineStageId}
\title{cppr_getPipelineStageId}
\usage{
cppr_getPipelineStageId(pipeline_id, name)
}
\arguments{
\item{pipeline_id}{numeric - the Copper Id of the pipeline you want information on}

\item{name}{string - the label (name) of the stage, e.g. "Quote sent"}
}
\value{
an numeric Id. The first matching Copper Id, or NULL if no match found or on error.
}
\description{
Get a Copper Id for the stage record identfied by the name argument
for use in the corresponding field in Opportunity records
}
\examples{
\dontrun{ 
pipes <- cppr_getPipelines() 
id <- cppr_getPipelineStageId(pipes$id[1],"Quote sent")
if(!is.null(id)) print(paste("stage 'Quote sent' has the id",id))
}
}
\seealso{
For details of data structures refer to \href{https://developer.copper.com/#getting-support}{Copper (TM) Developer API Guide}

Other Resource functions: \code{\link{cppr_extractCustomFieldValues}},
  \code{\link{cppr_getContactTypes}},
  \code{\link{cppr_getCustomFieldDatatype}},
  \code{\link{cppr_getCustomFieldDefinitions}},
  \code{\link{cppr_getCustomFieldId}},
  \code{\link{cppr_getCustomFieldName}},
  \code{\link{cppr_getCustomFieldOptionId}},
  \code{\link{cppr_getCustomFieldOptionName}},
  \code{\link{cppr_getCustomFieldOptionsAll}},
  \code{\link{cppr_getCustomerSourceId}},
  \code{\link{cppr_getCustomerSources}},
  \code{\link{cppr_getLeadStatusId}},
  \code{\link{cppr_getLeadStatuses}},
  \code{\link{cppr_getLossReasonId}},
  \code{\link{cppr_getLossReasons}},
  \code{\link{cppr_getPipelineStages}},
  \code{\link{cppr_getPipelines}}
}
