% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copperr_users.R
\name{cppr_getUserId}
\alias{cppr_getUserId}
\title{cppr_getUserId}
\usage{
cppr_getUserId(name)
}
\arguments{
\item{name}{the name of a Copper User : e.g. "Bob"}
}
\value{
a Copper ID, or NULL if not found
}
\description{
get the User Id for a named User
}
\examples{
\dontrun{ 
id <- cppr_getUserId("Bob")
}
}
\seealso{
Other user functions: \code{\link{cppr_getUsers}}
}
