% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copperr_init.R
\name{cppr_initialize}
\alias{cppr_initialize}
\title{cppr_initialize}
\usage{
cppr_initialize(uid, api_key, verbose = FALSE)
}
\arguments{
\item{uid}{string - the email address used for your PW account}

\item{api_key}{string - your API key from copper Settings>Preferences>API Keys}

\item{verbose}{boolean - set to TRUE if you are not getting the expected output 
and want to get verbose status messages printed to the console}
}
\value{
TRUE if successful
}
\description{
initializes the package by setting the fields of the private global environment.
}
\note{
You must call this function first before calling any other package function
}
\examples{
\dontrun{ 
cppr_initialize("foo@bar.com","1234567890ABCDEF")
}
}
