% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copperr_leads.R
\name{cppr_updateLead}
\alias{cppr_updateLead}
\title{cppr_updateLead}
\usage{
cppr_updateLead(id, list_of_fields)
}
\arguments{
\item{id}{numeric - the Id of the Lead to be updated}

\item{list_of_fields}{list - containing only the fields you want to modify}
}
\value{
the id of the Lead if update was successful, else NULL
}
\description{
Update an Lead record using the fields passed as arguments
}
\examples{
\dontrun{ 
fields <- list( `details` = "updated this Lead",
                `assignee_id` = myUserId )
if(cppr_updateLead(id,fields)) 
  sprintf("Successfully updated details of Lead \%d and reassigned to user \%d",id,myUserId)
}
}
\seealso{
\code{\link{cppr_getLeads}} for full list of Lead fields

Other Lead functions: \code{\link{cppr_createNewLead}},
  \code{\link{cppr_deleteLead}},
  \code{\link{cppr_getLeads}}
}
