% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copperr_opps.R
\name{cppr_updateOpp}
\alias{cppr_updateOpp}
\title{cppr_updateOpp}
\usage{
cppr_updateOpp(id, list_of_fields)
}
\arguments{
\item{id}{numeric - the Id of the Opportunity to be updated}

\item{list_of_fields}{list - containing only the fields you want to modify}
}
\value{
the id of the opp if update was successful, else NULL
}
\description{
Update an Opportunity record using the fields passed as arguments
}
\examples{
\dontrun{ 
fields <- list( `name` = "new name for this Opportunity",
                `assignee_id` = 12345 )
if(cppr_updateOpp(23456,fields)) 
  print("opportunity successfully updated")
}
}
\seealso{
\code{\link{cppr_getOppById}} for full list of Opportunity fields

For details of data structures refer to \href{https://developer.copper.com/#getting-support}{Copper (TM) Developer API Guide}

Other Opportunity functions: \code{\link{cppr_createNewOpp}},
  \code{\link{cppr_deleteOpp}},
  \code{\link{cppr_getOppById}},
  \code{\link{cppr_getOppByName}},
  \code{\link{cppr_getOpps}},
  \code{\link{cppr_updateOppCloseDate}},
  \code{\link{cppr_updateOppCustomFieldDirect}},
  \code{\link{cppr_updateOppCustomFieldDropdownMulti}},
  \code{\link{cppr_updateOppCustomFieldDropdown}},
  \code{\link{cppr_updateOppStage}}
}
