% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copperr_opps.R
\name{cppr_updateOppCloseDate}
\alias{cppr_updateOppCloseDate}
\title{cppr_updateOppCloseDate}
\usage{
cppr_updateOppCloseDate(id, closedate_sz)
}
\arguments{
\item{id}{numeric - the Id of the Opportunity to be updated}

\item{closedate_sz}{string - the close date in string format (e.g. "03/31/2018")}
}
\value{
the id of the opp if update was successful, else NULL
}
\description{
Update the 'Close Date' field of an Opportuntiy record
}
\note{
Unlike the other date fields used in Copper, the 'Close Date' field takes a 
text date, NOT a POSIX date class !!
}
\examples{
\dontrun{ 
if(cppr_updateOppCloseDate(12345,"03/31/2018")) 
  print("Close date changed successfully")
}
}
\seealso{
Other Opportunity functions: \code{\link{cppr_createNewOpp}},
  \code{\link{cppr_deleteOpp}},
  \code{\link{cppr_getOppById}},
  \code{\link{cppr_getOppByName}},
  \code{\link{cppr_getOpps}},
  \code{\link{cppr_updateOppCustomFieldDirect}},
  \code{\link{cppr_updateOppCustomFieldDropdownMulti}},
  \code{\link{cppr_updateOppCustomFieldDropdown}},
  \code{\link{cppr_updateOppStage}},
  \code{\link{cppr_updateOpp}}
}
