% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copperr_opps.R
\name{cppr_updateOppCustomFieldDropdown}
\alias{cppr_updateOppCustomFieldDropdown}
\title{cppr_updateOppCustomFieldDropdown}
\usage{
cppr_updateOppCustomFieldDropdown(id, field_name, value)
}
\arguments{
\item{id}{numeric - Copper Id for the Opportunity you are updating}

\item{field_name}{string - name of the custom field}

\item{value}{string - name of the dropdown option to find and set the id for}
}
\value{
the id of the opp if update was successful, else NULL
}
\description{
updates an Opportunity: changes the value for a custom field of dropdown type
}
\examples{
\dontrun{ 
if(cppr_updateOppCustomFieldDropdown(12345,"Industry","Finance")) 
  print("Opp updated successfully!")
}
}
\seealso{
For details of data structures refer to \href{https://developer.copper.com/#getting-support}{Copper (TM) Developer API Guide}

Other Opportunity functions: \code{\link{cppr_createNewOpp}},
  \code{\link{cppr_deleteOpp}},
  \code{\link{cppr_getOppById}},
  \code{\link{cppr_getOppByName}},
  \code{\link{cppr_getOpps}},
  \code{\link{cppr_updateOppCloseDate}},
  \code{\link{cppr_updateOppCustomFieldDirect}},
  \code{\link{cppr_updateOppCustomFieldDropdownMulti}},
  \code{\link{cppr_updateOppStage}},
  \code{\link{cppr_updateOpp}}
}
