% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copperr_opps.R
\name{cppr_updateOppCustomFieldDropdownMulti}
\alias{cppr_updateOppCustomFieldDropdownMulti}
\title{cppr_updateOppCustomFieldDropdownMulti}
\usage{
cppr_updateOppCustomFieldDropdownMulti(id, field_name, valueList)
}
\arguments{
\item{id}{numeric - Copper Id for the Opportunity you are updating}

\item{field_name}{string - name of custom field}

\item{valueList}{list - list of names of the dropdown options to select}
}
\value{
the id of the opp if update was successful, else NULL
}
\description{
updates an Opportunity: changes the value for a custom field of the Multi-choice dropdown type
You need to pass it the Copper Ids of the dropdown options, not the option names
}
\examples{
\dontrun{ 
optionsList <- list("Finance","Banking")
if(cppr_updateOppCustomFieldDropdown(12345,"Industry",optionsList)) 
print("Opp updated successfully!")
}
}
\seealso{
Other Opportunity functions: \code{\link{cppr_createNewOpp}},
  \code{\link{cppr_deleteOpp}},
  \code{\link{cppr_getOppById}},
  \code{\link{cppr_getOppByName}},
  \code{\link{cppr_getOpps}},
  \code{\link{cppr_updateOppCloseDate}},
  \code{\link{cppr_updateOppCustomFieldDirect}},
  \code{\link{cppr_updateOppCustomFieldDropdown}},
  \code{\link{cppr_updateOppStage}},
  \code{\link{cppr_updateOpp}}
}
