% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copperr_opps.R
\name{cppr_updateOppStage}
\alias{cppr_updateOppStage}
\title{cppr_updateOppStage
Update the 'Stage' field of an Opportuntiy record}
\usage{
cppr_updateOppStage(id, stage_id)
}
\arguments{
\item{id}{numeric - Copper Id for the Opportunity you are updating}

\item{stage_id}{numeric - the id of the Stage you want to move it to}
}
\value{
the id of the opp if update was successful, else NULL
}
\description{
cppr_updateOppStage
Update the 'Stage' field of an Opportuntiy record
}
\examples{
\dontrun{ 
stages <- cppr_getPipelineStages(12345)
newStage <- stages$id[stages$name=="Qualified"]
if(cppr_updateOppStage(12345,newStage)) 
  print("Stage changed successfully")
}
}
\seealso{
\code{\link{cppr_getPipelineStages}} to geta table of stage Ids and names

For details of data structures refer to \href{https://developer.copper.com/#getting-support}{Copper (TM) Developer API Guide}

Other Opportunity functions: \code{\link{cppr_createNewOpp}},
  \code{\link{cppr_deleteOpp}},
  \code{\link{cppr_getOppById}},
  \code{\link{cppr_getOppByName}},
  \code{\link{cppr_getOpps}},
  \code{\link{cppr_updateCompanyCustomFieldDirect}},
  \code{\link{cppr_updateOppCloseDate}},
  \code{\link{cppr_updateOppCustomFieldDirect}},
  \code{\link{cppr_updateOppCustomFieldDropdownMulti}},
  \code{\link{cppr_updateOppCustomFieldDropdown}},
  \code{\link{cppr_updateOpp}},
  \code{\link{cppr_updatePersonCustomFieldDirect}}
}
