% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copperr_people.R
\name{cppr_updatePerson}
\alias{cppr_updatePerson}
\title{cppr_updatePerson}
\usage{
cppr_updatePerson(id, list_of_fields)
}
\arguments{
\item{id}{numeric - the Id of the Person to be updated}

\item{list_of_fields}{list - a list of fields for the new Person 
\itemize{
  \item Field        Type    Details
  \item id           number  Unique identifier 
  \item name         string  firstname lastname
  \item address      address An encapsulation of the Person's street, city, state, postal code, and country.
  \item assignee_id  number Unique identifier of the User that will be the owner of the Person.
  \item etc
}}
}
\value{
the id of the Person if update was successful, else NULL
}
\description{
Update a Person record identified by the the id
}
\examples{
\dontrun{ 
if(!is.null(cppr_updatePerson(id))) 
  sprintf("Person \%d successfully removed",id)
}
}
\seealso{
For details of data structures refer to \href{https://developer.copper.com/#getting-support}{Copper (TM) Developer API Guide}

Other People functions: \code{\link{cppr_createNewPerson}},
  \code{\link{cppr_deletePerson}},
  \code{\link{cppr_getAllRecordsRelatedToPerson}},
  \code{\link{cppr_getOppsRelatedToPerson}},
  \code{\link{cppr_getPeople}},
  \code{\link{cppr_getPersonByEmail}},
  \code{\link{cppr_getPersonById}},
  \code{\link{cppr_getPersonByName}},
  \code{\link{cppr_getPersonId}}
}
