% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copperr_people.R
\name{cppr_updatePersonCustomFieldDirect}
\alias{cppr_updatePersonCustomFieldDirect}
\title{cppr_updatePersonCustomFieldDirect}
\usage{
cppr_updatePersonCustomFieldDirect(id, field_name, value)
}
\arguments{
\item{id}{numeric - Copper Id for the Person you are updating}

\item{field_name}{string - name of custom field, e.g. "Nickname"}

\item{value}{various types - value to set the field to. e.g. "Bucky" or 23 or TRUE}
}
\value{
the id of the opp if update was successful, else NULL
}
\description{
update any custom field of an Person that is not a dropdown (which needs an option id)
can be text, number, currency, date or checkbox
}
\examples{
\dontrun{ 
if(cppr_updatePersonCustomFieldDirect(202193, "Sales Region", "Mid-West"))
  print("Person updated successfully"))
}
}
\seealso{
Other Opportunity functions: \code{\link{cppr_createNewOpp}},
  \code{\link{cppr_deleteOpp}},
  \code{\link{cppr_getOppById}},
  \code{\link{cppr_getOppByName}},
  \code{\link{cppr_getOpps}},
  \code{\link{cppr_updateCompanyCustomFieldDirect}},
  \code{\link{cppr_updateOppCloseDate}},
  \code{\link{cppr_updateOppCustomFieldDirect}},
  \code{\link{cppr_updateOppCustomFieldDropdownMulti}},
  \code{\link{cppr_updateOppCustomFieldDropdown}},
  \code{\link{cppr_updateOppStage}},
  \code{\link{cppr_updateOpp}}
}
