% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copperr_init.R
\name{pkg_is_initialized}
\alias{pkg_is_initialized}
\title{pkg_is_initialized}
\usage{
pkg_is_initialized()
}
\arguments{
\item{x}{a value you are asserting is something}
}
\value{
TRUE if environment is initialized.
}
\description{
a test for use with the stopifnot() assertion 
a private function within this package
}
\keyword{internal}
