% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copperr_init.R
\name{printIfVerbose}
\alias{printIfVerbose}
\title{printIfVerbose}
\usage{
printIfVerbose(x, ...)
}
\arguments{
\item{x}{a value you are asserting is something

printIfVerbose("the value of x is",x)}
}
\description{
prints arguments if the verbose flag has been set to TRUE
a private function within this package
}
\keyword{internal}
