% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copperr_leads.R
\name{cppr_getLeads}
\alias{cppr_getLeads}
\title{cppr_getLeads}
\usage{
cppr_getLeads(assignee_id = NULL, country = NULL, state = NULL,
  tags = NULL, sort_by = "name", sort_direction = "asc")
}
\arguments{
\item{assignee_id}{numeric - select leads assigned to this User}

\item{country}{string - select leads from this country (specify the 2-letter country code, e.g. "NZ")}

\item{state}{string - select leads from this state (specify the full name of state, e.g. "Idaho")}

\item{tags}{string - select leads with one of these tags}

\item{sort_by}{string - "name", etc}

\item{sort_direction}{- allowed values "asc" or "desc"}
}
\value{
a data frame.\cr
Field          Type    Details\cr
id             number  Unique identifier \cr
name           string  firstname lastname\cr
address        data frame encapulsation of street,city,postcode,state, country\cr
assignee_id    number  Unique identifier of the User that will be the owner of the Lead.\cr
company_id     string  The unique identifier of the primary Company with which the Lead is associated.\cr
company_name   string  The name of the primary Company with which the Lead is associated.\cr
customer_source_id number  The unique identifier of the source of the Lead\cr
details        string  Description of the Lead.\cr
email[]       data frame  An array of email addresses belonging to the Lead.\cr
email[].email  string  An email address.\cr
email[].category string  The category of the email address.\cr
phone_numbers[] list  An array of phone numbers belonging to the Lead.\cr
phone_numbers[].number  string  A phone number.\cr
phone_numbers[].category  string  The category of the phone number.\cr
socials[]       list  An array of social profiles belonging to the Lead.\cr
socials[].url    string  The URL of a social profile.\cr
socials[].category  string  The category of the social profile.\cr
status         string_enum \cr
tags           list  An array of the tags associated with the Lead, represented as strings.\cr
title          string  The professional title of the Lead.\cr
websites[]     list  An array of websites belonging to the Lead.\cr
websites[].url  string  The URL of a website.\cr
websites[].category string  The category of the website.\cr
date_created   number  A Unix timestamp representing the time at which this Lead was created.\cr
date_modified  number  A Unix timestamp representing the time at which this Lead was last modified.\cr
custom_fields[] list  An array of custom field values belonging to the Person.\cr
custom_fields[]   \cr
.custom_field_definition_id number  The id of the Custom Field Definition for which this Custom Field stores a value.\cr
custom_fields[]   \cr
.value  mixed The value (number, string, option id, or timestamp) of this Custom Field.
}
\description{
Get a table containg all Leads from Copper. You can filter the recirds using the optional arguments.
Each row of the table is a Lead, represented by a list of field names and values. 
Some fields are themselves lists.
}
\examples{
\dontrun{ 
leads <- cppr_getLeads()
print(paste("retrieved",nrow(leads),"leads from PW"))
}
}
\seealso{
For details of data structures refer to \href{https://developer.copper.com/#getting-support}{Copper (TM) Developer API Guide}

Other Lead functions: \code{\link{cppr_createNewLead}},
  \code{\link{cppr_deleteLead}},
  \code{\link{cppr_updateLead}}
}
