% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copperr_people.R
\name{cppr_getOppsRelatedToPerson}
\alias{cppr_getOppsRelatedToPerson}
\title{cppr_getOppsRelatedToPerson}
\usage{
cppr_getOppsRelatedToPerson(id)
}
\arguments{
\item{id}{string - the id of the person}
}
\value{
a data frame of record ids , or NULL on error
Field        Type    Details
id           number  Unique identifier of the record
type         string  "opportunity"
}
\description{
get the ids of all opportunity records in Copper related to a Person.
}
\examples{
\dontrun{ 
relopps <- cppr_getOppsRelatedToPerson(12345)
print(paste("this person is related to",nrow(relopps),"opportunities"))
}
}
\seealso{
\code{\link{cppr_getAllRecordsRelatedToPerson}}

\code{\link{cppr_getPersonId}}

Other People functions: \code{\link{cppr_createNewPerson}},
  \code{\link{cppr_deletePerson}},
  \code{\link{cppr_getAllRecordsRelatedToPerson}},
  \code{\link{cppr_getPeople}},
  \code{\link{cppr_getPersonByEmail}},
  \code{\link{cppr_getPersonById}},
  \code{\link{cppr_getPersonByName}},
  \code{\link{cppr_getPersonId}},
  \code{\link{cppr_updatePerson}}
}
