% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copperr_people.R
\name{cppr_getPeople}
\alias{cppr_getPeople}
\title{cppr_getPeople}
\usage{
cppr_getPeople(assignee_id = NULL, country = NULL, state = NULL,
  tags = NULL, sort_by = "name", sort_direction = "asc")
}
\arguments{
\item{assignee_id}{numeric - select people assigned to this User}

\item{country}{string - select people from this country (specify the 2-letter country code, e.g. "NZ")}

\item{state}{string - select people from this state (specify the full name of state, e.g. "Idaho")}

\item{tags}{string - select people with one of these tags}

\item{sort_by}{string - sort records by "name", etc}

\item{sort_direction}{string - allowed values "asc" or "desc"}
}
\value{
a data frame, where each row is a Person, or NULL on error
Field          Type    Details
id             number  Unique identifier 
name           string  firstname lastname
address        address ecapulsation of street,city,postcode,state, country
assignee_id    number  Unique identifier of the User that will be the owner of the Person.
company_id     string  The unique identifier of the primary Company with which the Person is associated.
company_name   string  The name of the primary Company with which the Person is associated.
contact_type_id number  The unique identifier of the Contact Type of the Person.
details        string  Description of the Person.
emails[]       list  An array of email addresses belonging to the Person.
emails[].email  string  An email address.
emails[].category string  The category of the email address.
phone_numbers[] list  An array of phone numbers belonging to the Person.
phone_numbers[].number  string  A phone number.
phone_numbers[].category  string  The category of the phone number.
socials[]       list  An array of social profiles belonging to the Person.
socials[].url    string  The URL of a social profile.
socials[].category  string  The category of the social profile.
tags           list  An array of the tags associated with the Person, represented as strings.
title string  The professional title of the Person.
websites[]     list  An array of websites belonging to the Person.
websites[].url  string  The URL of a website.
websites[].category string  The category of the website.
date_created   number  A Unix timestamp representing the time at which this Person was created.
date_modified  number  A Unix timestamp representing the time at which this Person was last modified.
custom_fields[] list  An array of custom field values belonging to the Person.
custom_fields[]   
.custom_field_definition_id number  The id of the Custom Field Definition for which this Custom Field stores a value.
custom_fields[]   
.value  mixed The value (number, string, option id, or timestamp) of this Custom Field.
}
\description{
Get a table containg all People from Copper. You can filter the recirds using the optional arguments.
Each row of the table is a Person, represented by a list of field names and values. 
Some fields are themselves lists.
}
\examples{
\dontrun{ 
people <- cppr_getPeople()
print(paste("retrieved",nrow(people),"people from Copper"))
}
}
\seealso{
Other People functions: \code{\link{cppr_createNewPerson}},
  \code{\link{cppr_deletePerson}},
  \code{\link{cppr_getAllRecordsRelatedToPerson}},
  \code{\link{cppr_getOppsRelatedToPerson}},
  \code{\link{cppr_getPersonByEmail}},
  \code{\link{cppr_getPersonById}},
  \code{\link{cppr_getPersonByName}},
  \code{\link{cppr_getPersonId}},
  \code{\link{cppr_updatePerson}}
}
