% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copperr_people.R
\name{cppr_getPersonByEmail}
\alias{cppr_getPersonByEmail}
\title{cppr_getPersonByEmail}
\usage{
cppr_getPersonByEmail(email)
}
\arguments{
\item{email}{string - the Copper Id of the person}
}
\value{
a list, or NULL if unsuccessful
Field        Type    Details
id           number  Unique identifier 
name*        string  firstname lastname
address      address An encapsulation of the Person's street, city, state, postal code, and country.
assignee_id  number Unique identifier of the User that will be the owner of the Person.
etc
}
\description{
Get a People record identfied by the person's Name
}
\examples{
\dontrun{ 
person <- cppr_getPersonByEmail("john@foo.bar")
print(paste("his/her name is",person$name))
}
}
\seealso{
\code{\link{cppr_getPeople}} for full list of Person fields

Other People functions: \code{\link{cppr_createNewPerson}},
  \code{\link{cppr_deletePerson}},
  \code{\link{cppr_getAllRecordsRelatedToPerson}},
  \code{\link{cppr_getOppsRelatedToPerson}},
  \code{\link{cppr_getPeople}},
  \code{\link{cppr_getPersonById}},
  \code{\link{cppr_getPersonByName}},
  \code{\link{cppr_getPersonId}},
  \code{\link{cppr_updatePerson}}
}
